updated printing method
I changed the method I use to print such that I run 'jupyter nbconvert --to html [FILE].ipynb' after completion. Then I open the html file and print to a pdf. This results in no horizontal cutoffs in code, text, or plots, and it's easier to print without tweaking the scale.

I've deleted the old versions of my solutions for problem sets 1 and 2 and replaced them with the new method. Also standardised the name of the pset01 notebook from the old pset1 version to match the others